Reorganize agent packages
